requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
Django==4.2.1
django-bootstrap4==23.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
